Begin special-purpose DNS-SD wrapper API
The C FFI cannot pass in function pointers as far as I know,
which means it cannot call the core function we need.

Creating a more Idris-friendly bridging API makes interacting with C
far simpler, at the cost of having to write C.
It also looks like it will reduce the amount of FFI calls we'll have to
make, since most of the C function interaction will now happen from C.

Also: IO has infected 95% of our program at this point, due to the use of
FFI. That's the cost of interfacing with the real world, I guess.