Extract dnssd_bridge.h
This makes it far easier to review the public API of the bridge module.

I would have done this sooner (you could see earlier when I grouped things
towards the top of the .c file), but I worried I'd have to work out how
to use a Makefile and link in .o files as part of the IPKG / %include
configuration, and I didn't want to go there then.

Turns out, I didn't need to, so this was painless.