Make all functions in DNSSD total
The only problem child was collect', which has the potential to keep chasing
pointers forever. It's unlikely a host on the local network will ever have
more than a handful of records, so I just forcibly capped the number of
times it could recurse at 50. A more general solution would be to shift
to Data.CoList or Prelude.Stream and then guarantee productivity, but the
fuel argument approach is less invasive.